refactor: menu domain package move (tobe.domain -> domain)
