+++ src/main/java/kitchenpos/menus/domain/exception/InvalidMenuProductsPriceException.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain.exception;
+package kitchenpos.menus.domain.exception;
 
 public class InvalidMenuProductsPriceException extends IllegalArgumentException {
 }
