+++ src/main/java/kitchenpos/menus/domain/exception/NotFoundMenuGroupException.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain.exception;
+package kitchenpos.menus.domain.exception;
 
 import java.util.NoSuchElementException;
