+++ src/main/java/kitchenpos/menus/infra/jpa/JpaMenuGroupRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menus.infra.jpa;
 
-import kitchenpos.menus.tobe.domain.MenuGroup;
-import kitchenpos.menus.tobe.domain.MenuGroupRepository;
+import kitchenpos.menus.domain.MenuGroup;
+import kitchenpos.menus.domain.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
