+++ src/main/java/kitchenpos/menus/infra/jpa/JpaMenuRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menus.infra.jpa;
 
-import kitchenpos.menus.tobe.domain.Menu;
-import kitchenpos.menus.tobe.domain.MenuRepository;
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
